# Add AbbreviatedKey support to mergingIter