This PR implements the change discussed in #17, and also fixes a few bugs and comments.

The first commit just threads through the entire DB.Comparer to a bunch of places to grant access to the `AbbreviatedKey` function, the second commit actually adds support for making use of it to the mergingIter.

My conclusion is that this is a pessimization. Here's two runs of the `scan` benchmark (on my slow personal laptop):

before:
```
$ pebble scan bench --wipe
wiping bench
dir bench
concurrency 1
_elapsed_______rows/sec_______MB/sec_______ns/row
      1s      5691814.6         43.4        175.7
      2s      6126246.7         46.7        163.2
      3s      6272370.0         47.9        159.4
      4s      6256512.9         47.7        159.8
      5s      6279855.3         47.9        159.2
      6s      6248615.3         47.7        160.0
      7s      6274775.5         47.9        159.4
      8s      6294911.9         48.0        158.9
      9s      6214536.6         47.4        160.9
     10s      6255158.4         47.7        159.9

_elapsed___ops/sec(cum)__MB/sec(cum)__ns/row(avg)
   10.0s      6191469.5         47.2        161.5
```

after:
```
$ pebble scan bench --wipe
wiping bench
dir bench
concurrency 1
_elapsed_______rows/sec_______MB/sec_______ns/row
      1s      5332110.9         40.7        187.5
      2s      5766008.0         44.0        173.4
      3s      5833169.8         44.5        171.4
      4s      5885917.6         44.9        169.9
      5s      5902175.2         45.0        169.4
      6s      5891818.1         45.0        169.7
      7s      5905602.1         45.1        169.3
      8s      5876599.0         44.8        170.2
      9s      5807050.4         44.3        172.2
     10s      5837537.5         44.5        171.3

_elapsed___ops/sec(cum)__MB/sec(cum)__ns/row(avg)
   10.0s      5803635.5         44.3        172.3
```

Looking at the CPU profile, doing the comparisons are indeed faster, but the overhead of computing the abbreviated key outweighs its benefit.

My intuition says that a degradation in this case makes sense: each individual key is relatively short-lived in the mergingIter compared to batchskl (i.e., it will probably only have a handful of comparisons done against it), and so the overhead of computing the abbreviated key outweighs the benefit of having it around.

Posting this to get some more eyes on it to see if there's something I could be doing differently to get some improvements. Perhaps there's a better place to be computing the abbreviated keys.