@justinj 2019-03-05
Here's a benchstat for `BenchmarkMergingIterNext` changed to test 10, 20, 30, 40 iterators (which might be unrealistic given usual LSM height):
```
$ benchstat master.txt abbrev.txt
name                           old time/op  new time/op  delta
Justin/restart=16/count=10-24   168ns ± 0%   167ns ± 0%  -0.30%  (p=0.013 n=10+8)
Justin/restart=16/count=20-24   208ns ± 0%   204ns ± 0%  -1.57%  (p=0.000 n=10+9)
Justin/restart=16/count=30-24   237ns ± 1%   228ns ± 1%  -3.52%  (p=0.000 n=9+9)
Justin/restart=16/count=40-24   253ns ± 0%   246ns ± 1%  -2.64%  (p=0.000 n=7+9)
```

I also have a [sketchy patch](https://github.com/justinj/pebble/commit/0dd5a22e3825244bf1d8d4b0f8aff342f401b93d) to make computing the abbreviated key a little faster,

vs. abbreviated key:
```
$ benchstat abbrev.txt sketchy.txt
name                           old time/op  new time/op  delta
Justin/restart=16/count=10-24   167ns ± 0%   160ns ± 0%  -4.19%  (p=0.000 n=8+8)
Justin/restart=16/count=20-24   204ns ± 0%   197ns ± 0%  -3.48%  (p=0.000 n=9+9)
Justin/restart=16/count=30-24   228ns ± 1%   222ns ± 1%  -2.64%  (p=0.000 n=9+10)
Justin/restart=16/count=40-24   246ns ± 1%   242ns ± 0%  -1.91%  (p=0.000 n=9+8)
```

vs. no abbreviated key:
```
$ benchstat master.txt sketchy.txt
name                           old time/op  new time/op  delta
Justin/restart=16/count=10-24   168ns ± 0%   160ns ± 0%  -4.48%  (p=0.000 n=10+8)
Justin/restart=16/count=20-24   208ns ± 0%   197ns ± 0%  -4.99%  (p=0.000 n=10+9)
Justin/restart=16/count=30-24   237ns ± 1%   222ns ± 1%  -6.07%  (p=0.000 n=9+10)
Justin/restart=16/count=40-24   253ns ± 0%   242ns ± 0%  -4.50%  (p=0.000 n=7+8)
```