@petermattis 2019-03-05
> Here's a benchstat for BenchmarkMergingIterNext changed to test 10, 20, 30, 40 iterators (which might be unrealistic given usual LSM height):

I think the usual height of the LSM will be in the 6-7 level range. How was the data interleaved between the levels? In addition to the height, that will also make a significant difference (I think). 

> I also have a sketchy patch to make computing the abbreviated key a little faster,

That's not too sketchy, and it makes the change here much more compelling. I think unaligned reads only works on some architectures, but x86 is one of them. Radu is a good person to talk to about low-level bit twiddling like this.