@petermattis 2019-03-05
Another option for the "sketchy" fast-path for constructing an abbreviated key is to lift the code from Go runtime: `byteorder.LittleEndian.Uint64`. I believe that compiles down into a few assembly instructions.