@petermattis 2019-04-12
@justinj I took this PR for a spin trying to see why it isn't showing any perf diff. The first commit is actually a part of the problem. Using `Comparer.Compare` everywhere adds an indirection to hot code paths, vs just accessing `Compare` directly. Removing that reduced the delta at smaller numbers of levels, but didn't fully recover it. The win from using abbreviated keys only appears at higher numbers of levels than would be reasonably expected. I think it is safe to call this a failed experiment, close it and the associated issue, and move on to greener pastures.